requirements.txt
# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
tensorflow>=2.13.0

# Vision & Video Processing
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0
imageio>=2.31.0
av>=10.0.0

# Transformers & Pre-trained Models
transformers>=4.30.0
timm>=0.9.0
clip @ git+https://github.com/openai/CLIP.git

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
albumentations>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

# MLOps
wandb>=0.15.0
mlflow>=2.5.0
